move headers into include/mapbox folder - closes #99
