+++ test/our_variant_hello_world.cpp
@@ -1,4 +1,4 @@
-#include "variant.hpp"
+#include <mapbox/variant.hpp>
 
 #include <stdexcept>
