+++ test/t/binary_visitor_2.cpp
@@ -1,5 +1,5 @@
 
-#include "variant.hpp"
+#include <mapbox/variant.hpp>
 
 #define NAME_EXT " b-i-d"
 using variant_type = mapbox::util::variant<bool, int, double>;
