+++ test/t/binary_visitor_3.cpp
@@ -1,5 +1,5 @@
 
-#include "variant.hpp"
+#include <mapbox/variant.hpp>
 
 #define NAME_EXT " i-d-b"
 using variant_type = mapbox::util::variant<int, double, bool>;
