+++ test/t/binary_visitor_5.cpp
@@ -1,5 +1,5 @@
 
-#include "variant.hpp"
+#include <mapbox/variant.hpp>
 
 #define NAME_EXT " b-i-c-d-i"
 using variant_type = mapbox::util::variant<bool, int, char, double, int>;
