+++ test/t/variant.cpp
@@ -1,7 +1,7 @@
 #include "catch.hpp"
 
-#include "variant.hpp"
-#include "variant_io.hpp"
+#include <mapbox/variant.hpp>
+#include <mapbox/variant_io.hpp>
 
 #include <algorithm>
 #include <cstdint>
