+++ test/unique_ptr_test.cpp
@@ -8,7 +8,7 @@
 
 #include <boost/timer/timer.hpp>
 
-#include "variant.hpp"
+#include <mapbox/variant.hpp>
 
 using namespace mapbox;
